refactor: merge model-tiers into types to dissolve the import cycle (cycle-break #1)
`src/types.ts` imported `ModelTierSchema` from `src/model-tiers.ts`, while
`src/model-tiers.ts` imported `ProviderName` (type-only) back from `src/types.ts`.
That mutual import is a benign (type-only-at-runtime) cycle, but it blocks
cleanly extracting a standalone `@swarm/types` package in the monorepo split.

This commit dissolves the cycle by merging every export of `model-tiers.ts`
into `types.ts` under a dedicated `// ─── Model Tiers ───` section placed
directly after the imports, so `ModelTierSchema` (a runtime value) is defined
before its first use in this file. `ProviderName` is referenced only as a type
in the merged section, so the forward reference is safe (type refs are hoisted).
`src/model-tiers.ts` is removed and all importers are repointed to `./types`.

Behavior is byte-for-byte preserved: tier enum, default mappings,
`MODEL_TIER_<TIER>` / `MODEL_TIER_MAP` env+JSON overrides, legacy alias
normalization, and claim-time resolution are unchanged.

Prep for the `@swarm/types` monorepo package. No runtime behavior change.

Gates: tsc --noEmit clean, check-db-boundary.sh passes, biome check clean,
targeted tests (model-control, workflow-agent-task, http-api-integration,
migration-runner-regressions) 203 pass / 0 fail.
