+++ src/tests/model-control.test.ts
@@ -12,17 +12,17 @@ import {
   updateScheduledTask,
   upsertSwarmConfig,
 } from "../be/db";
-import {
-  parseModelTier,
-  resolveModelTier,
-  resolveTaskModelSelection,
-  splitLegacyModelAlias,
-} from "../model-tiers";
 import { runScheduleNow } from "../scheduler";
 import { createScheduleInputSchema } from "../tools/schedules/create-schedule";
 import { updateScheduleInputSchema } from "../tools/schedules/update-schedule";
 import { sendTaskInputSchema } from "../tools/send-task";
 import { taskActionInputSchema } from "../tools/task-action";
+import {
+  parseModelTier,
+  resolveModelTier,
+  resolveTaskModelSelection,
+  splitLegacyModelAlias,
+} from "../types";
 
 const TEST_DB_PATH = "./test-model-control.sqlite";
